Hrishi

+++ src/main/java/com/alpha/qspiderrestapi/service/impl/BranchServiceImpl.java
@@ -54,7 +54,7 @@ public class BranchServiceImpl implements BranchService {
 
 	@Autowired
 	private CityDao cityDao;
-	
+
 	@Autowired
 	private ValidatePhoneNumber validatePhoneNumber;
 
@@ -64,7 +64,7 @@ public class BranchServiceImpl implements BranchService {
 		branch.setBranchTitle(branch.getDisplayName() + "-" + branch.getBranchType());
 		for (String contact : branch.getContacts()) {
 			if (!validatePhoneNumber.isValidPhoneNumber(contact)) {
-				throw new InvalidPhoneNumberException("Invalid contact : "+contact);
+				throw new InvalidPhoneNumberException("Invalid contact : " + contact);
 			}
 		}
 		branch.setBranchFaqs(
@@ -154,9 +154,9 @@ public class BranchServiceImpl implements BranchService {
 						branch.setBranchName(branchView.getDisplayName());
 						branch.setBranchImage(branchView.getBranchImage());
 						branch.setLocation(branchView.getLocation());
-						branch.setPhoneNumber(branchView.getContacts());
-						branch.setUpcomingBatches(branchView.getUpcomingBatches());
-						branch.setOngoingBatches(branchView.getOngoingBatches());
+						branch.setPhoneNumber((branchView.getContacts()));
+						branch.setUpcomingBatches((long) branchView.getUpcomingBatches());
+						branch.setOngoingBatches((long) branchView.getOngoingBatches());
 						return branch;
 					}).sorted(Comparator.comparing(BranchDto::getBranchId)).collect(Collectors.toList());
 					course.setBranches(branches);
